Partially disentangle the clone code.
This still leaves some stuff dangling.
